requirement.txt
###### Requirements  ######

bcrypt
PyJWT
uvicorn
SQLAlchemy
fastapi